refactor: extract components and connect goals
