+++ src/components/GoalsOverview.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import List from "./List";
+
+export default function GoalsOverview(props) {
+  return <List title="Goals" type="goals" {...props} />;
+}
