+++ src/components/PeopleOverview.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import List from "./List";
+
+export default function PeopleOverview(props) {
+  return <List title="Tracked People Projects" type="people" {...props} />;
+}
