+++ src/components/ProgressOverview.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import List from "./List";
+
+export default function ProgressOverview(props) {
+  return <List title="Track Progress" type="progress" {...props} />;
+}
